# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MECHA API",
        "description": "API untuk Sistem Manajemen Bengkel Mobil",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/ai-diagnostics": {
            "get": {
                "tags": [
                    "AI Diagnostic"
                ],
                "summary": "Get semua AI diagnostics",
                "operationId": "381d642a016899c4497d8902483b1b61",
                "responses": {
                    "200": {
                        "description": "Data AI diagnostics retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "AI Diagnostic"
                ],
                "summary": "Buat AI diagnostic baru",
                "operationId": "d98f5e279e309cf71e557eac7efa0878",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "diagnostic_id",
                                    "symptoms"
                                ],
                                "properties": {
                                    "diagnostic_id": {
                                        "type": "integer"
                                    },
                                    "symptoms": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "AI Diagnostic created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register user baru",
                "operationId": "8a56853624e025573120a09a4c75d468",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Registrasi berhasil"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login berhasil"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "operationId": "fe8f3429cd6979b3b4517e186505f9f9",
                "responses": {
                    "200": {
                        "description": "Logout berhasil"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/profile": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get user profile",
                "operationId": "d32e56b6274eb68e980da6d91fbe9fd8",
                "responses": {
                    "200": {
                        "description": "Data user berhasil diambil"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/profile/update": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Update profile user",
                "operationId": "b8423619abf3ba75fe6142aed29a8eae",
                "responses": {
                    "200": {
                        "description": "Profile berhasil diupdate"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/change-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change password",
                "operationId": "540baf9b2324b156be71c57911de6714",
                "responses": {
                    "200": {
                        "description": "Password berhasil diubah"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/bengkels": {
            "get": {
                "tags": [
                    "Bengkel"
                ],
                "summary": "Get semua bengkel",
                "operationId": "798227e7b9a4af2a632de95f2cd9067a",
                "responses": {
                    "200": {
                        "description": "Data bengkels retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Bengkel"
                ],
                "summary": "Tambah bengkel baru",
                "operationId": "7f03be5f366260227f600f9f639ef025",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "address",
                                    "phone",
                                    "email",
                                    "city",
                                    "province",
                                    "status",
                                    "owner_name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "address": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "city": {
                                        "type": "string"
                                    },
                                    "province": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "aktif",
                                            "nonaktif"
                                        ]
                                    },
                                    "owner_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Bengkel created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/diagnostics": {
            "get": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Get semua diagnostics",
                "operationId": "6e7b767bc65c694579ba7584217e0c48",
                "responses": {
                    "200": {
                        "description": "Data diagnostics retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Diagnostic"
                ],
                "summary": "Tambah diagnostic baru",
                "operationId": "01615aeab16e9eadc3989c019c43e758",
                "responses": {
                    "201": {
                        "description": "Diagnostic created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/mechanics": {
            "get": {
                "tags": [
                    "Mechanic"
                ],
                "summary": "Get semua mechanics",
                "operationId": "69605d7ca31f7090676bce53082e6b53",
                "responses": {
                    "200": {
                        "description": "Data mechanics retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mechanic"
                ],
                "summary": "Tambah mechanic baru",
                "operationId": "d846cf8c8757bc42b9db33d614b84835",
                "responses": {
                    "201": {
                        "description": "Mechanic created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get semua orders",
                "operationId": "2cd7e03021e72d9d799c12dfe179c865",
                "responses": {
                    "200": {
                        "description": "Data orders retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Tambah order baru",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/order-items": {
            "get": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "Get semua order items",
                "operationId": "6a6b106e104d558bfc813be8f4adf2f4",
                "responses": {
                    "200": {
                        "description": "Data order items retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "OrderItem"
                ],
                "summary": "Tambah order item baru",
                "operationId": "3390f5f422c701c6e520cc66e9d18037",
                "responses": {
                    "201": {
                        "description": "Order item created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/ratings": {
            "get": {
                "tags": [
                    "Rating"
                ],
                "summary": "Get semua ratings",
                "operationId": "ad64604c65f061cafbb78a893548ad91",
                "responses": {
                    "200": {
                        "description": "Data ratings retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Rating"
                ],
                "summary": "Tambah rating baru",
                "operationId": "e876bdaebd2a715619d9d19e8be09398",
                "responses": {
                    "201": {
                        "description": "Rating created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/services": {
            "get": {
                "tags": [
                    "Service"
                ],
                "summary": "Get semua services",
                "operationId": "4c175c5d6594b4a5112bbe15a46e2bb4",
                "responses": {
                    "200": {
                        "description": "Data services retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Service"
                ],
                "summary": "Tambah service baru",
                "operationId": "b05b71c30267faa468172458580fb6fc",
                "responses": {
                    "201": {
                        "description": "Service created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/spare-parts": {
            "get": {
                "tags": [
                    "SparePart"
                ],
                "summary": "Get semua spare parts",
                "operationId": "ad3e96ab1c0eb723d55852782c5e77a5",
                "responses": {
                    "200": {
                        "description": "Data spare parts retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SparePart"
                ],
                "summary": "Tambah spare part baru",
                "operationId": "71f64eef3be34fc66b50c20dc4007bd8",
                "responses": {
                    "201": {
                        "description": "Spare part created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/transactions": {
            "get": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Get semua transactions",
                "operationId": "6581469356772e7a74a0d79897544f1a",
                "responses": {
                    "200": {
                        "description": "Data transactions retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Transaction"
                ],
                "summary": "Tambah transaction baru",
                "operationId": "ab765dced58cfe2f620929db31452145",
                "responses": {
                    "201": {
                        "description": "Transaction created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
        "/api/vehicles": {
            "get": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Get semua vehicles",
                "operationId": "727634867f59796290a2fff276979198",
                "responses": {
                    "200": {
                        "description": "Data vehicles retrieved successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehicle"
                ],
                "summary": "Tambah vehicle baru",
                "operationId": "a88e0bca4193c91e713d0441659d199c",
                "responses": {
                    "201": {
                        "description": "Vehicle created successfully"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "token": {
                "type": "http",
                "name": "Token",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}